backend/requirements_simple.txt
# VISTA Backend Requirements - Simplified Version
# Python 3.8+ required

# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3

# Database (MySQL only for now)
PyMySQL==1.1.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.0.5

# Authentication & Security
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Face Recognition
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
Pillow==10.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
pandas==2.1.1
openpyxl==3.1.2

# Development
pytest==7.4.3
pytest-flask==1.2.0
